requirements.txt
fastapi[standard]
pymongo
pydantic
requests
python-dotenv
